==> scripts/4-est_output_report.qmd <==
---
title: "ABC-XYZ Premise and Dataset Creation"
author: Jonah Lyon (jonah@a2-ai.com)
date: today
date-format: long
format:
  pdf:
    toc: true
    number-sections: true
    colorlinks: true
---

The purpose of this script is to use the imaginary patient population from the previous step to simulate out some "real" concentrations using the PopPK model under the proposed study design.  500 replicate simulations will be created so that in the next step, we can obtain a healthy range of estimates that can be considered for the given study design.  From the 500 estimates, % bias can be reported.


```{r, warning = FALSE, echo = FALSE, message = FALSE}
renv::snapshot()
```


```{r, echo = FALSE, warning = FALSE, message = FALSE}
library(tidyverse)
library(mrgsolve)
library(yspec)
library(bbr)

# data spec yaml file
spec <- ys_load("../data/derived/ppk.yaml")
derDir <- "../data/derived/dose_histories"
base_data <- read_csv(file.path(derDir, "subject_demog_dose.csv"), na = ".")

```

# Load PopPK Model Run for ABC-XYZ
```{r}
mod1 <- mread("../model/R-mrgsolve/1001.mod", 
              start = 0, end = 120, delta = 1)

```




```{r}
theme_set(theme_bw())

ref <- param(mod1) |> 
  as.data.frame()

simest_df <- param_estimates_batch("../model/nonmem/simest1") |> 
  mutate(
    sim = as.numeric(str_extract(run, "\\d+"))
  ) |> 
  arrange(sim) |> 
  select(sim, error_msg, THETA1, THETA2) |> 
  mutate(
    REFTHETA1 = ref$THETA1,
    REFTHETA2 = ref$THETA2,
    CLBIAS = (THETA1 - REFTHETA1)/REFTHETA1*100,
    VCBIAS = (THETA2 - REFTHETA2)/REFTHETA2*100
  ) |> 
  pivot_longer(cols = CLBIAS:VCBIAS, names_to = "PARAM", values_to = "RESULT") |> 
  mutate(
    PARAM = ifelse(PARAM == "VCBIAS", "Central Volume", "Clearance"),
    PARAM = as.factor(PARAM))
  

ggplot(simest_df, aes(x = PARAM, y = RESULT)) +
  geom_hline(aes(yintercept = 20), linetype = "dashed", col = "darkred") + 
  geom_hline(aes(yintercept = 0)) + 
  geom_hline(aes(yintercept = -20), linetype = "dashed", col = "darkred") + 
  geom_boxplot() +
  ggtitle("Bias of 500 Parameter Estimates with the corresponding sampling schedule") + 
  xlab("Parameter") + 
  ylab("Bias compared to original model (%)")

```
